Add feature tests for land and buildings transaction tax
There was already steps for the stamp duty in place, so
the only thing to do in the feature tests is to say to the tests
to look to the land and buildings transaction tax instead.

Removed the ftb locale from the lbtt tool because this tool
does not have first time buyers option. Probably it was a copy and paste.

After this the controllers, helpers and views could be refactored.