Add feature tests for land transaction tax for Wales
There was already steps for the stamp duty in place, so
the only thing to do in the feature tests is to say to the tests
to look to the land transaction tax page instead.

Removed the ftb locale from the tool because this tool
does not have first time buyers option.

After this the controllers, helpers and views could be refactored.