Refactor: Land Transaction Tax Calculator Wales (LTT)
This commit addresses Step 2 in the approach outlined in the first commit.

LTT controller inherits from the Transaction Tax Controller which was also used
as the 'parent' controller for the LBTT(Scotland) controller. Now that the LBTT
calculator is more akin to the Stamp Duty calculator, it makes more sense that the
LTT inherits from the application controller.

This commit makes the necessary changes to effect that change of parent controller. Specifically,

- add/amend controller methods to work better with views and translations
- make partials and views specific for the LTT model
- change folder structure to reflect LTT's "standalone" inheritance

Refer to pr 333, https://github.com/moneyadviceservice/mortgage_calculator/pull/333
which introduced the single template inheritance between LTT and LBTT