Refactor: Create parent calculator controller for SDLT and LBTT
This commit addresses Step 3 in the approach outlined in the first commit.

StampDuties and LandAndBuildingsTransactionTaxes controllers are now almost identical, as are the views.

This commit creates a parent controller for them to inherit from. It is a renaming of TransactionTaxesController with some additional code for the first time buyer journey.